requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.0
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
datasets==2.16.1
debugpy==1.8.0
decorator==5.1.1
deepsparse-nightly==1.7.0.20240131
defusedxml==0.7.1
diffusers==0.21.4
dill==0.3.7
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
fastjsonschema==2.19.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.46.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.10.0
future==0.18.3
geocoder==1.38.1
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.27.1
ipython==8.18.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
JPype1==1.5.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter-server-mathjax==0.2.6
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab_git==0.50.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
nbclient==0.9.0
nbconvert==7.12.0
nbdime==4.0.1
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
notebook_shim==0.2.3
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
onnx==1.14.1
onnxruntime==1.17.0
openai==1.12.0
optimum==1.13.2
optimum-deepsparse==0.1.0.dev1
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20221105
pdfplumber==0.10.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
plotly==5.18.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.25.2
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-machineid==0.5.1
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycparser==2.21
pycryptodome==3.20.0
pydantic==1.10.14
pydantic_core==2.16.2
Pygments==2.17.2
PyMuPDF==1.23.16
PyMuPDFb==1.23.9
pyparsing==3.1.1
PyPDF2==3.0.1
pypdfium2==4.26.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
ratelim==0.1.6
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.13.2
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
Send2Trash==1.8.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
sparsezoo-nightly==1.7.0.20240131
stack-data==0.6.3
sympy==1.12
tabula-py==2.9.0
tenacity==8.2.3
terminado==0.18.0
threadpoolctl==3.2.0
tiktoken==0.5.2
timm==0.9.12
tinycss2==1.2.1
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.36.2
triton==2.1.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
